+++ src/main/java/algoteg/Objetivo.java
@@ -4,6 +4,6 @@ public interface Objetivo {
 
     boolean cumplido();
     void setJugador(Jugador jugador);
-    String verObjetivo();
     void setMensajeObjetivo(String mensajeObjetivo);
+    String getMensajeObjetivo();
 }
